move more integration tests to e2e
