--- test/integration/master_routes_test.go
@@ -1,131 +1,0 @@
-package integration
-
-import (
-	"crypto/tls"
-	"crypto/x509"
-	"encoding/json"
-	"errors"
-	"io/ioutil"
-	"net/http"
-	"strings"
-	"testing"
-
-	knet "k8s.io/apimachinery/pkg/util/net"
-
-	testutil "github.com/openshift/origin/test/util"
-	testserver "github.com/openshift/origin/test/util/server"
-)
-
-func TestRootRedirect(t *testing.T) {
-	masterConfig, clusterAdminKubeConfig, err := testserver.StartTestMasterAPI()
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	defer testserver.CleanupMasterEtcd(t, masterConfig)
-
-	transport, err := anonymousHttpTransport(clusterAdminKubeConfig)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	req, err := http.NewRequest("GET", masterConfig.OAuthConfig.MasterPublicURL, nil)
-	req.Header.Set("Accept", "*/*")
-	resp, err := transport.RoundTrip(req)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("Expected %d, got %d", http.StatusOK, resp.StatusCode)
-	}
-	if resp.Header.Get("Content-Type") != "application/json" {
-		t.Fatalf("Expected %s, got %s", "application/json", resp.Header.Get("Content-Type"))
-	}
-	type result struct {
-		Paths []string
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		t.Fatalf("Unexpected error reading the body: %v", err)
-	}
-	var got result
-	json.Unmarshal(body, &got)
-}
-
-func TestWellKnownOAuth(t *testing.T) {
-	masterConfig, clusterAdminKubeConfig, err := testserver.StartTestMasterAPI()
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	defer testserver.CleanupMasterEtcd(t, masterConfig)
-
-	transport, err := anonymousHttpTransport(clusterAdminKubeConfig)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	req, err := http.NewRequest("GET", masterConfig.OAuthConfig.MasterPublicURL+"/.well-known/oauth-authorization-server", nil)
-	req.Header.Set("Accept", "*/*")
-	resp, err := transport.RoundTrip(req)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("Expected %d, got %d", http.StatusOK, resp.StatusCode)
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		t.Fatalf("Unexpected error reading the body: %v", err)
-	}
-	if !strings.Contains(string(body), "authorization_endpoint") {
-		t.Fatal("Expected \"authorization_endpoint\" in the body.")
-	}
-}
-
-func TestWellKnownOAuthOff(t *testing.T) {
-	masterConfig, err := testserver.DefaultMasterOptions()
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	defer testserver.CleanupMasterEtcd(t, masterConfig)
-	masterConfig.OAuthConfig = nil
-	clusterAdminKubeConfig, err := testserver.StartConfiguredMasterAPI(masterConfig)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	clusterAdminClientConfig, err := testutil.GetClusterAdminClientConfig(clusterAdminKubeConfig)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	transport, err := anonymousHttpTransport(clusterAdminKubeConfig)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	req, err := http.NewRequest("GET", clusterAdminClientConfig.Host+"/.well-known/oauth-authorization-server", nil)
-	req.Header.Set("Accept", "*/*")
-	resp, err := transport.RoundTrip(req)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	if resp.StatusCode != http.StatusNotFound {
-		t.Fatalf("Expected %d, got %d", http.StatusNotFound, resp.StatusCode)
-	}
-}
-
-func anonymousHttpTransport(clusterAdminKubeConfig string) (*http.Transport, error) {
-	restConfig, err := testutil.GetClusterAdminClientConfig(clusterAdminKubeConfig)
-	if err != nil {
-		return nil, err
-	}
-	pool := x509.NewCertPool()
-	if ok := pool.AppendCertsFromPEM(restConfig.TLSClientConfig.CAData); !ok {
-		return nil, errors.New("failed to add server CA certificates to client pool")
-	}
-	return knet.SetTransportDefaults(&http.Transport{
-		TLSClientConfig: &tls.Config{
-			// only use RootCAs from client config, especially no client certs
-			RootCAs: pool,
-		},
-	}), nil
-}
